mantis-backend/pyproject.toml
[project]
name = "mantis-backend"
version = "1.0.0"
description = "M.A.N.T.I.S. — Market Analysis & Network Tactical Integration System Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "openai>=1.60.0",
    "anthropic>=0.40.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "aiosqlite>=0.20.0",
    "structlog>=24.0.0",
    "web3>=7.0.0",
]

[project.scripts]
mantis = "mantis.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/mantis"]
